backend/requirements.base.txt
bcrypt==3.1.1
cffi==1.8.3
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pycparser==2.16
pymongo==3.3.0
requests==2.11.1
six==1.10.0
Werkzeug==0.11.11
PyYAML==3.12
configparser==3.5.0
enum34==1.1.6
python-pam==1.8.2
